+++ lib/repo_tx_registration.go
@@ -13,32 +13,35 @@ import (
 // True if the actor is ultimately, definitively registered for the specified topic, else False
 // Idempotent in registration
 func (node *NodeConfig) RegisterWorkerIdempotently(ctx context.Context, config WorkerConfig) (bool, error) {
+	log := log.With().Str("rpc", node.Wallet.NodeRpc).Uint64("topicId", config.TopicId).Str("actorType", "worker").Logger()
+	log.Info().Msg("Registering worker")
+
 	isRegistered, err := node.IsWorkerRegistered(ctx, config.TopicId)
 	if err != nil {
 		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the node is already registered for topic as worker, skipping")
 		return false, err
 	}
 	if isRegistered {
-		log.Info().Str("rpc", node.Wallet.NodeRpc).Uint64("topicId", config.TopicId).Msg("Worker node already registered for topic")
+		log.Info().Msg("Already registered for topic")
 		return true, nil
 	} else {
-		log.Info().Str("rpc", node.Wallet.NodeRpc).Uint64("topicId", config.TopicId).Msg("Worker node not yet registered for topic. Attempting registration...")
+		log.Info().Msg("Node not yet registered. Attempting registration...")
 	}
 
 	moduleParams, err := node.Chain.EmissionsQueryClient.GetParams(ctx, &emissionstypes.GetParamsRequest{})
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not get chain params for worker ")
+		log.Error().Err(err).Msg("Could not get chain params for node")
 		return false, err
 	}
 
 	balance, err := node.GetBalance(ctx)
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the worker node has enough balance to register, skipping")
+		log.Error().Err(err).Msg("Could not check if the node has enough balance to register, skipping")
 		return false, err
 	}
 	if !balance.GTE(moduleParams.Params.RegistrationFee) {
-		log.Error().Str("rpc", node.Wallet.NodeRpc).Str("balance", balance.String()).Msg("Worker node does not have enough balance to register, skipping.")
-		return false, fmt.Errorf("worker node does not have enough balance to register, skipping")
+		log.Error().Str("balance", balance.String()).Msg("Node does not have enough balance to register, skipping.")
+		return false, fmt.Errorf("Node does not have enough balance to register, skipping")
 	}
 
 	msg := &emissionstypes.RegisterRequest{
@@ -47,26 +50,26 @@ func (node *NodeConfig) RegisterWorkerIdempotently(ctx context.Context, config W
 		Owner:     node.Chain.Address,
 		IsReputer: false,
 	}
-	res, err := node.SendDataWithRetry(ctx, msg, "Register worker node", 0)
+	res, err := node.SendDataWithRetry(ctx, msg, "Register node", 0)
 	if err != nil {
 
 		txHash := ""
 		if res != nil {
 			txHash = res.TxHash
 		}
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not register the worker node with the Allora blockchain")
+		log.Error().Err(err).Str("txHash", txHash).Msg("Could not register the node with the Allora blockchain")
 		return false, err
 	}
 
 	// Give time for the tx to be included in a block
 	log.Debug().Int64("delay", node.Wallet.RetryDelay).Msg("Waiting to check registration status to be included in a block...")
 	if DoneOrWait(ctx, node.Wallet.RetryDelay) {
-		log.Error().Err(ctx.Err()).Str("rpc", node.Wallet.NodeRpc).Msg("Waiting to check registration status failed")
+		log.Error().Err(ctx.Err()).Msg("Waiting to check registration status failed")
 		return false, ctx.Err()
 	}
 	isRegistered, err = node.IsWorkerRegistered(ctx, config.TopicId)
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the node is already registered for topic as worker, skipping")
+		log.Error().Err(err).Msg("Could not check if the node is already registered for topic, skipping")
 		return false, err
 	}
 
@@ -77,30 +80,33 @@ func (node *NodeConfig) RegisterWorkerIdempotently(ctx context.Context, config W
 // Actor may be either a worker or a reputer
 // Idempotent in registration and stake addition
 func (node *NodeConfig) RegisterAndStakeReputerIdempotently(ctx context.Context, config ReputerConfig) (bool, error) {
+	log := log.With().Str("rpc", node.Wallet.NodeRpc).Uint64("topicId", config.TopicId).Str("actorType", "reputer").Logger()
+	log.Info().Msg("Registering reputer")
+
 	isRegistered, err := node.IsReputerRegistered(ctx, config.TopicId)
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the node is already registered for topic as reputer, skipping")
+		log.Error().Err(err).Msg("Could not check if the node is already registered for topic as reputer, skipping")
 		return false, err
 	}
 
 	if isRegistered {
-		log.Info().Uint64("topicId", config.TopicId).Msg("Reputer node already registered")
+		log.Info().Msg("Already registered")
 	} else {
-		log.Info().Uint64("topicId", config.TopicId).Msg("Reputer node not yet registered. Attempting registration...")
+		log.Info().Msg("Node not yet registered. Attempting registration...")
 
 		balance, err := node.GetBalance(ctx)
 		if err != nil {
-			log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the Reputer node has enough balance to register, skipping")
+			log.Error().Err(err).Msg("Could not check if the node has enough balance to register, skipping")
 			return false, err
 		}
 		moduleParams, err := node.Chain.EmissionsQueryClient.GetParams(ctx, &emissionstypes.GetParamsRequest{})
 		if err != nil {
-			log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not get chain params for reputer")
+			log.Error().Err(err).Msg("Could not get chain params for reputer")
 			return false, err
 		}
 		if !balance.GTE(moduleParams.Params.RegistrationFee) {
-			log.Error().Str("rpc", node.Wallet.NodeRpc).Msg("Reputer node does not have enough balance to register, skipping.")
-			return false, fmt.Errorf("reputer node does not have enough balance to register, skipping")
+			log.Error().Msg("Node does not have enough balance to register, skipping.")
+			return false, fmt.Errorf("Node does not have enough balance to register, skipping")
 		}
 
 		msgRegister := &emissionstypes.RegisterRequest{
@@ -109,46 +115,46 @@ func (node *NodeConfig) RegisterAndStakeReputerIdempotently(ctx context.Context,
 			Owner:     node.Chain.Address,
 			IsReputer: true,
 		}
-		res, err := node.SendDataWithRetry(ctx, msgRegister, "Register reputer node", 0)
+		res, err := node.SendDataWithRetry(ctx, msgRegister, "Register node", 0)
 		if err != nil {
 
 			txHash := ""
 			if res != nil {
 				txHash = res.TxHash
 			}
-			log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not register the reputer node with the Allora blockchain")
+			log.Error().Err(err).Str("txHash", txHash).Msg("Could not register the node with the Allora blockchain")
 			return false, err
 		}
 
 		// Give time for the tx to be included in a block
 		log.Debug().Int64("delay", node.Wallet.RetryDelay).Msg("Waiting to check registration status to be included in a block...")
 		if DoneOrWait(ctx, node.Wallet.RetryDelay) {
-			log.Error().Err(ctx.Err()).Str("rpc", node.Wallet.NodeRpc).Msg("Waiting to check registration status failed")
+			log.Error().Err(ctx.Err()).Msg("Waiting to check registration status failed")
 			return false, ctx.Err()
 		}
 		isRegistered, err = node.IsReputerRegistered(ctx, config.TopicId)
 		if err != nil {
-			log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the node is already registered for topic as reputer, skipping")
+			log.Error().Err(err).Msg("Could not check if the node is already registered for topic, skipping")
 			return false, err
 		}
 		if !isRegistered {
-			log.Error().Str("rpc", node.Wallet.NodeRpc).Uint64("topicId", config.TopicId).Msg("Reputer node not registered after all retries")
-			return false, fmt.Errorf("reputer node not registered after all retries")
+			log.Error().Msg("Node not registered after all retries")
+			return false, fmt.Errorf("Node not registered after all retries")
 		}
 	}
 
 	stake, err := node.GetReputerStakeInTopic(ctx, config.TopicId, node.Chain.Address)
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the reputer node has enough balance to stake, skipping")
+		log.Error().Err(err).Msg("Could not check if the node has enough balance to stake, skipping")
 		return false, err
 	}
 
 	minStake := cosmossdk_io_math.NewInt(config.MinStake)
 	if minStake.LTE(stake) {
-		log.Info().Str("rpc", node.Wallet.NodeRpc).Msg("Reputer stake above minimum requested stake, skipping adding stake.")
+		log.Info().Msg("Stake above minimum requested stake, skipping adding stake.")
 		return true, nil
 	} else {
-		log.Info().Str("rpc", node.Wallet.NodeRpc).Interface("stake", stake).Interface("minStake", minStake).Interface("stakeToAdd", minStake.Sub(stake)).Msg("Reputer stake below minimum requested stake, adding stake.")
+		log.Info().Interface("stake", stake).Interface("minStake", minStake).Interface("stakeToAdd", minStake.Sub(stake)).Msg("Stake below minimum requested stake, adding stake.")
 	}
 
 	msgAddStake := &emissionstypes.AddStakeRequest{
@@ -156,30 +162,30 @@ func (node *NodeConfig) RegisterAndStakeReputerIdempotently(ctx context.Context,
 		Amount:  minStake.Sub(stake),
 		TopicId: config.TopicId,
 	}
-	res, err := node.SendDataWithRetry(ctx, msgAddStake, "Add reputer stake", 0)
+	res, err := node.SendDataWithRetry(ctx, msgAddStake, "Add stake", 0)
 	if err != nil {
 		txHash := ""
 		if res != nil {
 			txHash = res.TxHash
 		}
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not stake the reputer node with the Allora blockchain in specified topic")
+		log.Error().Err(err).Str("txHash", txHash).Msg("Could not stake the node with the Allora blockchain in specified topic")
 		return false, err
 	}
 
 	// Give time for the tx to be included in a block
 	log.Debug().Int64("delay", node.Wallet.RetryDelay).Msg("Waiting to check stake status to be included in a block...")
 	if DoneOrWait(ctx, node.Wallet.RetryDelay) {
-		log.Error().Err(ctx.Err()).Str("rpc", node.Wallet.NodeRpc).Msg("Waiting to check stake status failed")
+		log.Error().Err(ctx.Err()).Msg("Waiting to check stake status failed")
 		return false, ctx.Err()
 	}
 	stake, err = node.GetReputerStakeInTopic(ctx, config.TopicId, node.Chain.Address)
 	if err != nil {
-		log.Error().Err(err).Str("rpc", node.Wallet.NodeRpc).Msg("Could not check if the reputer node has enough balance to stake, skipping")
+		log.Error().Err(err).Msg("Could not check if the node has enough balance to stake, skipping")
 		return false, err
 	}
 	if stake.LT(minStake) {
-		log.Error().Str("rpc", node.Wallet.NodeRpc).Interface("stake", stake).Interface("minStake", minStake).Msg("Reputer stake below minimum requested stake, skipping.")
-		return false, fmt.Errorf("reputer stake below minimum requested stake, skipping")
+		log.Error().Interface("stake", stake).Interface("minStake", minStake).Msg("Stake below minimum requested stake, skipping.")
+		return false, fmt.Errorf("Stake below minimum requested stake, skipping")
 	}
 
 	return true, nil